Refactor skill tests to adventure tests
